Hammer imported old code into shape to comply with storage-ng standards
